+++ MicroserviceGet/src/main/java/com/example/demo/reps/CompanyRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.models.Company;
 
-public interface CompanyRepository extends JpaRepository<Company, Long>{
+public interface CompanyRepository extends JpaRepository<Company, Integer>{
 
 	Collection<Company> findByNameContainingIgnoreCase(String name);
 }
