+++ MicroserviceGet/src/main/java/com/example/demo/reps/EducationRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.models.Education;
 
 
-public interface EducationRepository extends JpaRepository<Education,Long>{
+public interface EducationRepository extends JpaRepository<Education,Integer>{
 	Collection<Education> findByLevelOfEducationContainingIgnoreCase(String levelOfEducation);
 	Collection<Education> findByNameContainingIgnoreCase(String name);
