+++ MicroserviceGet/src/main/java/com/example/demo/reps/SectorRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.models.Company;
 import com.example.demo.models.Sector;;
 
-public interface SectorRepository extends JpaRepository<Sector,Long>{
+public interface SectorRepository extends JpaRepository<Sector,Integer>{
 
 	Collection<Sector> findByNameContainingIgnoreCase(String name);
 	Collection<Sector> findByCompany(Company c);
